New feature: jilter sru_z3950 handles x-target=host .
Filter sru_z3950 now handles x-target=host . The value (host) is
the address of a target (vhost). The vhost can be used by, for
example, z3950_client to connect to any Z39.50 server. Note that
at this stage the database is NOT set by the x-target. This must
also be done in order for this feature to be useful.