Allow control of pass-through of vhosts for virt_db
A new element in virt_db 'pass-vhosts' has a boolean value. If
true, it the virt_db passes vhost-enabled sessions - that is leaving
them untouched. This was also the old behavior of MP. A value of
false will make virt_db always inspect a Z39.50 session regardless
of whether vhost was set or not. This is new and default behavior.