Retrieval transform operational
Settings useTurboMarc, elementSet, recordEncoding, requestSyntax,
transform are in use. They are currently ONLY in use if
record syntax is XML and element set is "pz2" and requestSyntax
is defined (in torus config). In other cases, the zoom filter
is more or less pass through with Z39.50 settings taking effect
(syntax, elementSetName).