Filter frontend_net may log statistics
If a custom message element in frontend_net configuration it makes
the frontend_net filter produce a log entry whenever an operation
is completed. The log is always written using yaz_log with the
custom message given, the package ID (session ID) , the time
the operation has been active (end-to-end) , the threads currently
active/total and the size of the request input/output queues.